Improve code formatting, import extension imports and cleanup.
