--- .idea/modules/velocity-language-kotlin.main.iml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="FacetManager">
-    <facet type="minecraft" name="Minecraft">
-      <configuration>
-        <autoDetectTypes>
-          <platformType>ADVENTURE</platformType>
-        </autoDetectTypes>
-      </configuration>
-    </facet>
-  </component>
-</module>
