+++ src/main/kotlin/com/velocitypowered/api/kt/event/EventManagerExt.kt
@@ -1,0 +1,46 @@
+package com.velocitypowered.api.kt.event
+
+import com.velocitypowered.api.event.Event
+import com.velocitypowered.api.event.EventManager
+import com.velocitypowered.api.event.EventTask
+import com.velocitypowered.api.event.PostOrder
+import kotlin.coroutines.Continuation
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
+import kotlin.coroutines.startCoroutine
+
+/**
+ * Registers an event listener for the event [E] for the given [plugin]. The listener will use a
+ * suspended coroutine, allowing you to use a coroutine context to process the event in a
+ * non-blocking way.
+ */
+inline fun <reified E : Event> EventManager.on(
+  plugin: Any, order: Short = PostOrder.NORMAL, crossinline handler: suspend (E) -> Unit
+) =
+  register(plugin, E::class.java, order) { event ->
+    suspended {
+      handler(event)
+    }
+  }
+
+/**
+ * Marks the specified function as a suspended function, which uses the event continuation system in
+ * Velocity to allow you to process the event in a non-blocking way.
+ */
+fun suspended(fn: suspend () -> Unit): EventTask {
+  return EventTask.withContinuation { continuation ->
+    val completion = object : Continuation<Unit> {
+      override val context: CoroutineContext
+        get() = EmptyCoroutineContext
+
+      override fun resumeWith(result: Result<Unit>) {
+        if (result.isFailure) {
+          continuation.resumeWithException(result.exceptionOrNull())
+        } else {
+          continuation.resume()
+        }
+      }
+    }
+    fn.startCoroutine(completion)
+  }
+}
