+++ src/main/kotlin/com/velocitypowered/api/kt/proxy/connection/PlayerExt.kt
@@ -1,4 +1,4 @@
-package com.velocitypowered.api.language.kotlin.ext
+package com.velocitypowered.api.kt.proxy.connection
 
 import com.velocitypowered.api.proxy.connection.Player
 import com.velocitypowered.api.proxy.connection.ServerConnection
@@ -6,37 +6,37 @@ import com.velocitypowered.api.proxy.player.ClientSettings
 import com.velocitypowered.api.proxy.player.TabList
 import com.velocitypowered.api.util.GameProfile
 import com.velocitypowered.api.util.ModInfo
-import java.util.*
+import java.util.UUID
 
 val Player.uniqueId: UUID
-    inline get() = id()
+  inline get() = id()
 
 val Player.username: String
-    inline get() = username()
+  inline get() = username()
 
 val Player.connectedServer: ServerConnection?
-    inline get() = connectedServer()
+  inline get() = connectedServer()
 
 val Player.clientSettings: ClientSettings
-    inline get() = clientSettings()
+  inline get() = clientSettings()
 
 val Player.modInfo: ModInfo?
-    inline get() = modInfo()
+  inline get() = modInfo()
 
 val Player.ping: Long
-    inline get() = ping()
+  inline get() = ping()
 
 val Player.onlineMode: Boolean
-    inline get() = onlineMode()
+  inline get() = onlineMode()
 
 var Player.gameProfileProperties: Collection<GameProfile.Property>
-    inline get() = gameProfile().properties()
-    inline set(value) {
-        setGameProfileProperties(value.toList())
-    }
+  inline get() = gameProfile().properties()
+  inline set(value) {
+    setGameProfileProperties(value.toList())
+  }
 
 val Player.gameProfile: GameProfile
-    inline get() = gameProfile()
+  inline get() = gameProfile()
 
 val Player.tabList: TabList
-    inline get() = tabList()
+  inline get() = tabList()
